+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
-
+## 30.0.11
+New:
+- `GeoCalendar`: Added 'isToDateDisabled' and 'isFromDateDisabled' props to force to select only one side of the range
+- `GeoCalendarDropdown`: Added 'isToDateDisabled' and 'isFromDateDisabled' props to force to select only one side of the range
 ## 30.0.10
 Changed:
 - Update to node 14